pyproject.toml
[project]
name = "glavkrepezh"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiosmtplib>=5.1.0",
    "alembic>=1.18.3",
    "asyncpg>=0.31.0",
    "fastapi[standard]>=0.128.0",
    "itsdangerous>=2.2.0",
    "openpyxl>=3.1.5",
    "pandas>=3.0.0",
    "passlib[bcrypt]>=1.7.4",
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.22",
    "python-slugify>=8.0.4",
    "slowapi>=0.1.9",
    "sqlalchemy>=2.0.46",
    "starlette-admin>=0.16.0",
    "uvicorn>=0.40.0",
]
